feat: fixing the Cant column that it was a string

+++ src/app.py
@@ -62,7 +62,8 @@ def _clean_df(df:pd) -> pd.DataFrame:
             df["Description"].iloc[u] = df["Description"].iloc[u]+ "".join(strings_cant[:len(strings_cant)-1])
         if len(strings_des) >= 2:
             df["Description"].loc[u] = "".join(strings_des[1:])
-    
+
+    df["Cant"] = df["Cant"].astype(int)
     pivot =  df.groupby(["Barcode_Registry", "Description"], as_index=False).agg({"Cant": "sum"})
     return pivot
